introduce schedulerInterface and cron
`schedulerInterface` represents the actions that a backend needs
to satisfy to be used as a scheduler. At the moment, we only
have cron.
